requirements.txt
# 项目依赖示例
# 在部署或本地运行前，使用 pip install -r requirements.txt
# openai  # 如果使用 OpenAI API，可取消注释并指定版本
# python-dotenv  # 如果使用 .env 加载环境变量
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
